+++ NEWS.md
@@ -4,8 +4,8 @@
 
 * New function `detect()`.
 * New function `dictionary_sugar()`.
-* It is now asserted that the return value of Dictionaries is an R6 object.
-* Fix some more assertions in `Dictionary`
+* It is now asserted that the return value of `Dictionary$get()` is an R6 object.
+* Fix some assertions in `Dictionary`
 
 ## v0.1.2